+++ rust/src/structs/value.rs
@@ -1,6 +1,8 @@
 use crate::sieve_ir_generated::sieve_ir as g;
 use crate::Result;
 use flatbuffers::{FlatBufferBuilder, ForwardsUOffset, Vector, WIPOffset};
+use num_bigint_dig;
+use num_bigint_dig::prime::probably_prime;
 
 /// A Value is a field element encoded least-significant-byte-first (little-endian). Trailing zeros may be omitted.
 ///
@@ -44,4 +46,18 @@ pub fn build_values_vector<'bldr: 'args, 'args: 'mut_bldr, 'mut_bldr>(
         .map(|g_value| build_value(builder, g_value))
         .collect();
     builder.create_vector(&g_values)
+}
+
+pub fn is_probably_prime(value: &Value) -> bool {
+    let value = num_bigint_dig::BigUint::from_bytes_le(value);
+    probably_prime(&value, 10)
+}
+
+#[test]
+fn test_is_probably_prime() {
+    let value: Value = vec![187]; // 187=11*17
+    assert!(!is_probably_prime(&value));
+
+    let value: Value = vec![101];
+    assert!(is_probably_prime(&value));
 }